fix: resolve compilation errors and warnings in CUDA allocator, scheduler, and Llama model

+++ crates/vllm-engine/src/sampling.rs
@@ -21,8 +21,8 @@ impl Sampler {
         
         let processor = LogitsProcessor::new(
             seed,
-            Some(params.temperature),
-            Some(params.top_p),
+            Some(params.temperature as f64),
+            Some(params.top_p as f64),
         );
 
         Self { processor }
